fix fabrics, remove comments
