+++ alymova.ksenia/T1/inputProcess.hpp
@@ -7,10 +7,6 @@
 namespace alymova
 {
   void makeShape(std::istream& in, CompositeShape& shapes, bool& wrong_shape_flag, double* scale_params);
-  Shape* makeRectangle(std::istream& in);
-  Shape* makeCircle(std::istream& in);
-  Shape* makeRegular(std::istream& in);
-  void readParameters(std::istream& in, double* params, size_t size);
   void print(std::ostream& out, const Shape* shape);
   void print(std::ostream& out, const CompositeShape& shapes);
   void clear(Shape** shapes);
